==> 1. Scraping_WatchCount.ipynb <==
# Vinyl Grading Project
This project aims to construct an open source package to classify vinyl according to the [Goldmine Grading](https://www.goldminemag.com/collector-resources/record-grading-101).

First we need to create a Dataset. We will scrap some pages in search of labeled audio.

## Scrapin WatchCount

WatchCount is an archive of ebay sold items. I've selected one user that uses a template for the description and that includes always an audio. This will help us to create semi-automatically the dataset.

# install required packages if needed
!pip install pandas selenium requests

# Imports
import pandas as pd
from selenium import webdriver
from selenium.webdriver.common.by import By

# Let's open a session
driver = webdriver.Firefox()
url = "http://www.watchcount.com/completed.php?bfw=1&bslr=wills-rare-records"
driver.get(url)



Let's create a dataframe for the url to the actual item on ebay and the title of the item sold.

df_links = pd.DataFrame(columns=["Link","Title"])

df_links

Now is time to scrape all the table of sold items from this seller.

# This is the table's xpath
table = driver.find_element(By.XPATH, "/html/body/div/table/tbody/tr[3]/td/table/tbody/tr/td[2]/div/table[2]/tbody/tr[2]/td/div/table[1]")

# Find all rows within the table
rows = table.find_elements(By.TAG_NAME, "tr")

# Loop through each row
for row in rows:
    # Find all link elements within the row
    links = row.find_elements(By.TAG_NAME, "a")
    # Loop through each link element
    if len(links)==11:

    # Extract the URL from the href attribute
        url = links[1].get_attribute("href")
        text = links[1].get_attribute("text")
        
        df_links.loc[len(df_links)] = [url, text]


Let's have a look on the dataframe of the first scraped page.

df_links

Look's good. Let's scrape the rest of 8 pages.

for i in range(1,9):
    driver = webdriver.Firefox()
    url = "http://www.watchcount.com/completed.php?bslr=wills-rare-records&csbin=all&cssrt=ts&bfw=1&bpg={}#serp".format(i)
    driver.get(url)
    table = driver.find_element(By.XPATH, "/html/body/div/table/tbody/tr[3]/td/table/tbody/tr/td[2]/div/table[2]/tbody/tr[2]/td/div/table[1]")

    # Find all rows within the table
    rows = table.find_elements(By.TAG_NAME, "tr")

    # Loop through each row
    print(i)
    for row in rows:
        try:
            # Find all link elements within the row
            links = row.find_elements(By.TAG_NAME, "a")
            
            # Loop through each link element
            if len(links)==11:

                # Extract the URL from the href attribute
                url = links[1].get_attribute("href")
                text = links[1].get_attribute("text")
                df_links.loc[len(df_links)] = [url, text]
        except:
            print(row)
driver.quit()

Let's store the output.

df_links.to_csv("./output/wills.csv", index = False)
