==> 5. Dataset Creation.ipynb <==
# Dataset Creation and Preprocessing

Now that we have the dataframe and the audio clips of the Side A, we need to preprocess the data and create the Dataset to training our model.

## Preprocess dataframe
We will add some info as length of the audio and the route to the *Side A* audio

import pandas as pd
import librosa
import librosa.display
import os



wills = pd.read_csv("./output/wills_audio_SIDEA.csv")
wills.head()

As there were some errors, let's search for the audios in the Side A folder.

wills["SIDEA_path"] = wills["mp3_path"].apply(lambda x: f"./output/audio/SIDEA/{os.path.basename(x)}")
wills["SIDEA_path"]

wills.SIDEA_path.isna().value_counts()

wills = wills.loc[wills.SIDEA_path.notna()]

wills["CHECK_SIDEA"] = wills["SIDEA_path"].apply(lambda x: os.path.exists(x))

wills.CHECK_SIDEA.value_counts()

We remove them

wills = wills.loc[wills["CHECK_SIDEA"]==True]
wills.drop(columns=["CHECK_SIDEA"], inplace = True)

wills

wills.drop(columns = ["condition"], inplace = True)
wills.rename(columns={"A_first_part" : "condition"}, inplace = True)

## Dataset Creation

Finally we create the dataset.

from datasets import Dataset
dataset = Dataset.from_dict({"audio": wills["SIDEA_path"].tolist(), "condition":wills["condition"].tolist()})

dataset.save_to_disk("./output/dataset")
